Rename course registration and course view module.
