+++ app/build.gradle.kts
@@ -104,8 +104,8 @@ dependencies {
     implementation(project(":core:device"))
     implementation(project(":core:websocket"))
 
-    implementation(project(":feature:course_registration"))
-    implementation(project(":feature:course_view"))
+    implementation(project(":feature:course-registration"))
+    implementation(project(":feature:course-view"))
     implementation(project(":feature:dashboard"))
     implementation(project(":feature:login"))
     implementation(project(":feature:exercise_view"))
@@ -122,8 +122,8 @@ dependencies {
     kover(project(":core:ui"))
     kover(project(":core:device"))
     kover(project(":core:websocket"))
-    kover(project(":feature:course_registration"))
-    kover(project(":feature:course_view"))
+    kover(project(":feature:course-registration"))
+    kover(project(":feature:course-view"))
     kover(project(":feature:dashboard"))
     kover(project(":feature:login"))
     kover(project(":feature:exercise_view"))
