+++ feature/course-registration/build.gradle.kts
@@ -1,0 +1,17 @@
+plugins {
+    id("artemis.android.feature")
+    id("artemis.android.library.compose")
+}
+
+android {
+    namespace = "de.tum.informatics.www1.artemis.native_app.feature.courseregistration"
+}
+
+dependencies {
+    implementation(project(":core:ui"))
+    implementation(project(":core:data"))
+    implementation(project(":core:device"))
+
+    testImplementation(project(":feature:login"))
+    testImplementation(project(":feature:login-test"))
+}
