+++ feature/course-registration/src/main/res/values/course_registration_ui_strings.xml
