+++ feature/course-view/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/courseview/course_view_module.kt
@@ -1,0 +1,19 @@
+package de.tum.informatics.www1.artemis.native_app.feature.courseview
+
+import de.tum.informatics.www1.artemis.native_app.feature.courseview.ui.CourseViewModel
+import org.koin.androidx.viewmodel.dsl.viewModel
+import org.koin.dsl.module
+
+val courseViewModule = module {
+    viewModel {
+        CourseViewModel(
+            it.get(),
+            get(),
+            get(),
+            get(),
+            get(),
+            get(),
+            get()
+        )
+    }
+}
