+++ feature/course-view/src/main/res/values/course_ui_strings.xml
