--- feature/course_registration/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/course_registration/course_registration_module.kt
@@ -1,11 +1,0 @@
-package de.tum.informatics.www1.artemis.native_app.feature.course_registration
-
-import de.tum.informatics.www1.artemis.native_app.feature.course_registration.service.CourseRegistrationService
-import de.tum.informatics.www1.artemis.native_app.feature.course_registration.service.impl.CourseRegistrationServiceImpl
-import org.koin.androidx.viewmodel.dsl.viewModel
-import org.koin.dsl.module
-
-val courseRegistrationModule = module {
-    single<CourseRegistrationService> { CourseRegistrationServiceImpl(get()) }
-    viewModel { RegisterForCourseViewModel(get(), get(), get(), get()) }
-}
