--- feature/course_view/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/course_view/GroupedByWeek.kt
@@ -1,30 +1,0 @@
-package de.tum.informatics.www1.artemis.native_app.feature.course_view
-
-import kotlinx.datetime.LocalDate
-
-/**
- * Data points grouped by a single week
- */
-internal sealed class GroupedByWeek<T>(val items: List<T>) {
-
-    /**
-     * Key which can be used to to identify the instance in a map
-     */
-    abstract val key: String
-
-    class BoundToWeek<T>(
-        val firstDayOfWeek: LocalDate,
-        val lastDayOfWeek: LocalDate,
-        items: List<T>
-    ) : GroupedByWeek<T>(items) {
-        override val key: String
-            get() = firstDayOfWeek.toString()
-    }
-
-    class Unbound<T>(
-        items: List<T>
-    ) : GroupedByWeek<T>(items) {
-        override val key: String
-            get() = "Unbound"
-    }
-}
