--- feature/course_view/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/course_view/course_view_module.kt
@@ -1,19 +1,0 @@
-package de.tum.informatics.www1.artemis.native_app.feature.course_view
-
-import de.tum.informatics.www1.artemis.native_app.feature.course_view.ui.CourseViewModel
-import org.koin.androidx.viewmodel.dsl.viewModel
-import org.koin.dsl.module
-
-val courseViewModule = module {
-    viewModel {
-        CourseViewModel(
-            it.get(),
-            get(),
-            get(),
-            get(),
-            get(),
-            get(),
-            get()
-        )
-    }
-}
