+++ settings.gradle.kts
@@ -29,9 +29,9 @@ include(":core:websocket")
 include(":core:device")
 include(":core:device-test")
 include(":core:core-test")
-include(":feature:course_registration")
+include(":feature:course-registration")
 include(":feature:dashboard")
-include(":feature:course_view")
+include(":feature:course-view")
 include(":feature:login")
 include(":feature:login-test")
 include(":feature:exercise_view")
